[KAN-70] rename: 로그인 회원가입 submit 함수 이름 변경
